+++ inst/examples/ex-coordinates.R
@@ -11,5 +11,8 @@ head(get_coordinates(X, margin = 1, principal = TRUE))
 head(get_coordinates(X, margin = 1, principal = FALSE))
 
 ## Tidy principal coordinates
-head(tidy(X, margin = 1, axes = c(1, 2)))
-head(tidy(X, margin = 2, axes = c(1, 2)))
+head(tidy(X, margin = 1))
+head(tidy(X, margin = 2))
+
+head(augment(X, margin = 1, axes = c(1, 2)))
+head(augment(X, margin = 2, axes = c(1, 2)))